refactor: untangle two module boundaries
getEnvironmentVariable lived in lib/streams/streamAuction, so messageBuilder
imported it across a content -> streams edge for a function with nothing to do
with auction streaming. Move it to lib/helpers/env. Two parser tests had
mocked the whole streamAuction module just to stub this helper; they now mock
lib/helpers/env directly and no longer depend on streaming internals they do
not exercise.

The handler modules used `module.exports = fn`, forcing index.ts to load them
with untyped require() calls. Export named functions and import them, so a
rename fails the build rather than at boot - the same class of failure as the
handler scan this branch already removed.
