+++ src/handlers/Event.ts
@@ -2,7 +2,7 @@ import { Client } from 'discord.js';
 import { join } from 'path';
 import { loadEvents } from './eventLoader';
 
-module.exports = (client: Client) => {
+export function registerEventHandlers(client: Client) {
 	const eventsDir = join(__dirname, '../events');
 	loadEvents(eventsDir, client);
-};
+}
