+++ src/lib/helpers/env.ts
@@ -1,0 +1,7 @@
+export function getEnvironmentVariable(name: string): string {
+	const value = process.env[name];
+	if (!value) {
+		throw new Error(`Environment variable ${name} is not defined.`);
+	}
+	return value;
+}
