+++ src/lib/streams/streamAuction.ts
@@ -6,14 +6,7 @@ import {
 	packEmbedsForDiscord,
 } from '../content/messages/messageBuilder';
 import { gracefullyHandleError } from '../helpers/errors';
-
-export function getEnvironmentVariable(name: string): string {
-	const value = process.env[name];
-	if (!value) {
-		throw new Error(`Environment variable ${name} is not defined.`);
-	}
-	return value;
-}
+import { getEnvironmentVariable } from '../helpers/env';
 
 export type AuctionData = {
 	buying: ItemType[];
